refactor: move metrics to their own subpackage
