+++ scripts/reevaluate.py
@@ -13,8 +13,8 @@ from timeeval.adapters.docker import SCORES_FILE_NAME as DOCKER_SCORES_FILE_NAME
 from timeeval.constants import RESULTS_CSV, HYPER_PARAMETERS, METRICS_CSV, ANOMALY_SCORES_TS
 from timeeval.data_types import ExecutionType
 from timeeval.core.experiments import Experiment as TimeEvalExperiment
+from timeeval.metrics import FScoreAtK, PrecisionAtK
 from timeeval.utils.datasets import load_labels_only
-from timeeval.utils.metrics import PrecisionAtK, FScoreAtK
 
 # required to build a lookup-table for algorithm implementations
 import timeeval_experiments.algorithms as algorithms
