+++ timeeval/__init__.py
@@ -22,7 +22,7 @@ from ._version import __version__
 from .algorithm import Algorithm
 from .data_types import AlgorithmParameter, TrainingType, InputDimensionality
 from .datasets import Datasets, DatasetManager, MultiDatasetManager
+from .metrics import Metric, DefaultMetrics
 from .remote_configuration import RemoteConfiguration
 from .resource_constraints import ResourceConstraints
 from .timeeval import TimeEval, Status
-from .utils.metrics import Metric, DefaultMetrics
